requirements.txt
PySide6
pyqt6
python-dotenv
opencv-python
pytesseract
numpy
